+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.3
 - BUGFIX: rescue @ssh.close when far end closes disgracefully (ALU ISAM)
 - FEATURE: Alcatel-Lucent ISAM 7302/7330 model added by @jalmargyyk
+- FEATURE: Huawei VRP model added by @jalmargyyk
 
 # 0.2.2
 - BUGFIX: mark node as failure if unknown error is raised